Handle Form Response Submit

+++ src/server/helpers/trigger.js
@@ -6,9 +6,13 @@ export const adjustFormSubmitTrigger = () => {
   const form = FormApp.getActiveForm()
   const triggers = ScriptApp.getUserTriggers(form)
 
+  const documentProperties = PropertiesService.getDocumentProperties()
+  const triggerNeeded = documentProperties.getProperty('configuration') !== null
+
   // Create a new trigger if required; delete existing trigger
   //   if it is not needed.
   let existingTrigger = null
+
   for (let i = 0; i < triggers.length; i += 1) {
     if (triggers[i].getEventType() === ScriptApp.EventType.ON_FORM_SUBMIT) {
       existingTrigger = triggers[i]
@@ -16,13 +20,12 @@ export const adjustFormSubmitTrigger = () => {
     }
   }
 
-  // TODO: Optimize this later, at the moment it's for cleaning up trigger in case of error
-  if (existingTrigger) {
+  if (triggerNeeded && !existingTrigger) {
+    ScriptApp.newTrigger('respondToFormSubmit')
+      .forForm(form)
+      .onFormSubmit()
+      .create()
+  } else if (!triggerNeeded && existingTrigger) {
     ScriptApp.deleteTrigger(existingTrigger)
   }
-
-  ScriptApp.newTrigger('respondToFormSubmit')
-    .forForm(form)
-    .onFormSubmit()
-    .create()
 }
